views: Split views/auth.py out of core views file.
