+++ zproject/backends.py
@@ -127,7 +127,7 @@ class SocialAuthMixin(ZulipAuthMixin):
         # type: (UserProfile, *Any, **Any) -> Optional[HttpResponse]
         # This function needs to be imported from here due to the cyclic
         # dependency.
-        from zerver.views import login_or_register_remote_user
+        from zerver.views.auth import login_or_register_remote_user
 
         return_data = kwargs.get('return_data', {})
